+++ saleor/graphql/decorators.py
@@ -5,6 +5,8 @@ from typing import Iterable, Union
 from graphql_jwt import exceptions
 from graphql_jwt.decorators import context
 
+from ..core.permissions import AccountPermissions
+
 
 def account_passes_test(test_func):
     """Determine if user/app has permission to access to content."""
@@ -26,8 +28,11 @@ def _permission_required(perms: Iterable[Enum], context):
     if context.user.has_perms(perms):
         return True
     app = getattr(context, "app", None)
-    if app and app.has_perms(perms):
-        return True
+    if app:
+        # for now MANAGE_STAFF permission for app is not supported
+        if AccountPermissions.MANAGE_STAFF in perms:
+            return False
+        return app.has_perms(perms)
     return False